+++ src/main/java/io/xlorey/FluxLoader/utils/patch/PatchChatServer.java
@@ -37,17 +37,17 @@ public class PatchChatServer extends PatchFile{
             if (targetNode != null) {
                 InsnList eventInvoker = new InsnList();
 
-                eventInvoker.add(new InsnNode(Opcodes.ICONST_2)); // Размер массива
+                eventInvoker.add(new InsnNode(Opcodes.ICONST_2));
                 eventInvoker.add(new TypeInsnNode(Opcodes.ANEWARRAY, "java/lang/Object"));
 
                 eventInvoker.add(new InsnNode(Opcodes.DUP));
                 eventInvoker.add(new InsnNode(Opcodes.ICONST_0));
-                eventInvoker.add(new VarInsnNode(Opcodes.ALOAD, 4)); // Загружаем var4
+                eventInvoker.add(new VarInsnNode(Opcodes.ALOAD, 4));
                 eventInvoker.add(new InsnNode(Opcodes.AASTORE));
 
                 eventInvoker.add(new InsnNode(Opcodes.DUP));
                 eventInvoker.add(new InsnNode(Opcodes.ICONST_1));
-                eventInvoker.add(new VarInsnNode(Opcodes.ALOAD, 5)); // Загружаем var5
+                eventInvoker.add(new VarInsnNode(Opcodes.ALOAD, 5));
                 eventInvoker.add(new InsnNode(Opcodes.AASTORE));
 
                 eventInvoker.add(new LdcInsnNode("onChatServerMessage"));