+++ src/main/java/io/xlorey/FluxLoader/utils/patch/PatchGameServer.java
@@ -1,8 +1,10 @@
 package io.xlorey.FluxLoader.utils.patch;
 
+import io.xlorey.FluxLoader.utils.Logger;
 import org.objectweb.asm.Opcodes;
 import org.objectweb.asm.Type;
 import org.objectweb.asm.tree.*;
+import zombie.network.packets.DeadPlayerPacket;
 
 
 /**
@@ -82,17 +84,84 @@ public class PatchGameServer extends PatchFile{
             }
 
         });
+        PatchTools.injectEventInvokerWithLocalField(filePath, "receivePlayerDeath", 3, "onPlayerDeath", new int[] {3}, true);
+        PatchTools.injectEventInvokerWithLocalField(filePath, "receiveZombieDeath", 3, "onZombieDeath", new int[] {3}, true);
+        PatchTools.injectEventInvokerWithLocalField(filePath, "receiveReceiveCommand", 3, "onSendChatCommand", new int[] {1, 3}, true);
+        PatchTools.injectIntoClass(filePath, "handleServerCommand", true, method -> {
+            InsnList toInject = new InsnList();
 
-        PatchTools.injectIntoClass(filePath, "receivePlayerConnect", true, method -> {
-            Type[] argumentTypes = Type.getArgumentTypes(method.desc);
+            toInject.add(new VarInsnNode(Opcodes.ALOAD, 1));
+            LabelNode labelIfNotNull = new LabelNode();
+            toInject.add(new JumpInsnNode(Opcodes.IFNONNULL, labelIfNotNull));
 
-            if (argumentTypes.length >= 3 && argumentTypes[2].getDescriptor().equals("Ljava/lang/String;")) {
-                InsnList eventInvoker = PatchTools.createEventInvokerInsnList("onPlayerConnect", argumentTypes, true);
-                method.instructions.insertBefore(method.instructions.getFirst(), eventInvoker);
+            toInject.add(new LdcInsnNode("onSendConsoleCommand"));
+            toInject.add(new InsnNode(Opcodes.ICONST_1));
+            toInject.add(new TypeInsnNode(Opcodes.ANEWARRAY, "java/lang/Object"));
+            toInject.add(new InsnNode(Opcodes.DUP));
+            toInject.add(new InsnNode(Opcodes.ICONST_0));
+            toInject.add(new VarInsnNode(Opcodes.ALOAD, 0));
+            toInject.add(new InsnNode(Opcodes.AASTORE));
+            toInject.add(new MethodInsnNode(
+                    Opcodes.INVOKESTATIC,
+                    "io/xlorey/FluxLoader/shared/EventManager",
+                    "invokeEvent",
+                    "(Ljava/lang/String;[Ljava/lang/Object;)V",
+                    false
+            ));
+
+            toInject.add(new VarInsnNode(Opcodes.ALOAD, 0));
+            toInject.add(new MethodInsnNode(
+                    Opcodes.INVOKESTATIC,
+                    "io/xlorey/FluxLoader/shared/CommandsManager",
+                    "handleCustomConsoleCommand",
+                    "(Ljava/lang/String;)Ljava/lang/String;",
+                    false
+            ));
+
+            toInject.add(new InsnNode(Opcodes.DUP));
+            LabelNode labelIfResultNotNull = new LabelNode();
+            toInject.add(new JumpInsnNode(Opcodes.IFNULL, labelIfResultNotNull));
+
+            toInject.add(new InsnNode(Opcodes.ARETURN));
+
+            toInject.add(labelIfResultNotNull);
+            toInject.add(new InsnNode(Opcodes.POP));
+
+            toInject.add(labelIfNotNull);
+
+            method.instructions.insert(toInject);
+        });
+        PatchTools.injectIntoClass(filePath, "receiveReceiveCommand", true, method -> {
+            AbstractInsnNode targetNode = null;
+            for (AbstractInsnNode currentNode = method.instructions.getFirst(); currentNode != null; currentNode = currentNode.getNext()) {
+                if (currentNode instanceof VarInsnNode) {
+                    VarInsnNode varInsnNode = (VarInsnNode) currentNode;
+                    if (varInsnNode.getOpcode() == Opcodes.ALOAD && varInsnNode.var == 4) {
+                        AbstractInsnNode checkNode = currentNode.getNext();
+                        if (checkNode instanceof JumpInsnNode && checkNode.getOpcode() == Opcodes.IFNONNULL) {
+                            targetNode = currentNode;
+                            break;
+                        }
+                    }
+                }
+            }
+
+            if (targetNode != null) {
+                InsnList toInject = new InsnList();
+                toInject.add(new VarInsnNode(Opcodes.ALOAD, 1));
+                toInject.add(new VarInsnNode(Opcodes.ALOAD, 3));
+                toInject.add(new MethodInsnNode(
+                        Opcodes.INVOKESTATIC,
+                        "io/xlorey/FluxLoader/shared/CommandsManager",
+                        "handleCustomChatCommand",
+                        "(Lzombie/core/raknet/UdpConnection;Ljava/lang/String;)Ljava/lang/String;",
+                        false
+                ));
+                toInject.add(new VarInsnNode(Opcodes.ASTORE, 4));
+
+                method.instructions.insertBefore(targetNode, toInject);
             }
         });
-
-
         PatchTools.injectEventInvoker(filePath, "disconnectPlayer", "onPlayerDisconnect", true);
         PatchTools.injectEventInvoker(filePath, "addIncoming", "onAddIncoming", true);
     }